fix: merged code generation service
